# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAnonymi API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/documents": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Create Pasted Draft Route",
        "operationId": "create_pasted_draft_route_api_v1_documents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedDraftView"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents/{document_id}/handoff": {
      "get": {
        "tags": [
          "team review"
        ],
        "summary": "Get Handoff",
        "operationId": "get_handoff_api_v1_documents__document_id__handoff_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "team review"
        ],
        "summary": "Put Handoff",
        "operationId": "put_handoff_api_v1_documents__document_id__handoff_put",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoffInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/teammates": {
      "get": {
        "tags": [
          "team review"
        ],
        "summary": "Get Teammates",
        "operationId": "get_teammates_api_v1_documents__document_id__teammates_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeammateView"
                  },
                  "title": "Response Get Teammates Api V1 Documents  Document Id  Teammates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/approval": {
      "post": {
        "tags": [
          "team review"
        ],
        "summary": "Approval",
        "operationId": "approval_api_v1_documents__document_id__approval_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/findings/{finding_id}/comments": {
      "get": {
        "tags": [
          "team review"
        ],
        "summary": "Comments",
        "operationId": "comments_api_v1_documents__document_id__findings__finding_id__comments_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Finding Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentView"
                  },
                  "title": "Response Comments Api V1 Documents  Document Id  Findings  Finding Id  Comments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "team review"
        ],
        "summary": "Comment",
        "operationId": "comment_api_v1_documents__document_id__findings__finding_id__comments_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Finding Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/comments/{comment_id}": {
      "delete": {
        "tags": [
          "team review"
        ],
        "summary": "Remove",
        "operationId": "remove_api_v1_documents__document_id__comments__comment_id__delete",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Comment Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalInput": {
        "properties": {
          "expected": {
            "$ref": "#/components/schemas/VersionRef"
          },
          "confirmed_preview": {
            "type": "boolean",
            "title": "Confirmed Preview"
          }
        },
        "type": "object",
        "required": [
          "expected",
          "confirmed_preview"
        ],
        "title": "ApprovalInput"
      },
...
      "CommentInput": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "expected": {
            "$ref": "#/components/schemas/VersionRef"
          },
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "expected",
          "text"
        ],
        "title": "CommentInput"
      },
      "CommentView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "finding_id": {
            "type": "string",
            "format": "uuid",
            "title": "Finding Id"
          },
          "author_email": {
            "type": "string",
            "title": "Author Email"
          },
          "is_mine": {
            "type": "boolean",
            "title": "Is Mine"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "finding_id",
          "author_email",
          "is_mine",
          "text",
          "created_at"
        ],
        "title": "CommentView"
      },
...
      "CreateDraftRequest": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/FindingCategory"
            },
            "type": "array",
            "title": "Categories"
          },
          "phone_region": {
            "type": "string",
            "maxLength": 2,
            "minLength": 2,
            "title": "Phone Region",
            "default": "PH"
          },
          "retention_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 30.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Retention Days"
          },
          "preset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Id"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "source"
        ],
        "title": "CreateDraftRequest"
      },
...
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "draft",
          "scanning",
          "needs_review",
          "ready",
          "exported",
          "failed",
          "expired",
          "deleted"
        ],
        "title": "DocumentStatus"
      },
...
      "ErrorDetail": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "issue": {
            "type": "string",
            "title": "Issue"
          }
        },
        "type": "object",
        "required": [
          "field",
          "issue"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "FindingCategory": {
        "type": "string",
        "enum": [
          "person",
          "organization",
          "location",
          "address",
          "identifier",
          "custom",
          "email",
          "phone"
        ],
        "title": "FindingCategory"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HandoffInput": {
        "properties": {
          "expected": {
            "$ref": "#/components/schemas/VersionRef"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "require_approval": {
            "type": "boolean",
            "title": "Require Approval"
          }
        },
        "type": "object",
        "required": [
          "expected",
          "reviewer_id",
          "require_approval"
        ],
        "title": "HandoffInput"
      },
      "HandoffView": {
        "properties": {
          "version": {
            "$ref": "#/components/schemas/VersionRef"
          },
          "is_owner": {
            "type": "boolean",
            "title": "Is Owner"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "reviewer_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Email"
          },
          "reviewer_active": {
            "type": "boolean",
            "title": "Reviewer Active"
          },
          "require_approval": {
            "type": "boolean",
            "title": "Require Approval"
          },
          "approved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "generation": {
            "type": "integer",
            "title": "Generation"
          }
        },
        "type": "object",
        "required": [
          "version",
          "is_owner",
          "reviewer_id",
          "reviewer_email",
          "reviewer_active",
          "require_approval",
          "approved_at",
          "generation"
        ],
        "title": "HandoffView"
      },
...
      "SavedDraftView": {
        "properties": {
          "version": {
            "$ref": "#/components/schemas/VersionRef"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          }
        },
        "type": "object",
        "required": [
          "version",
          "expires_at",
          "status"
        ],
        "title": "SavedDraftView"
      },
...
      "TeammateView": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "email"
        ],
        "title": "TeammateView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VersionRef": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "source_revision_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Revision Id"
          },
          "decision_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Decision Version"
          },
          "settings_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Settings Version"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "source_revision_id",
          "decision_version",
          "settings_version"
        ],
        "title": "VersionRef"
      },
...
    }
  }
}